Added institute details retrieval function taking a bank code instead of a BIC.
This function is more reliable if you have a bank code as it queries the country specific classes instead of relying on data from the ECB (which doesn't include bank codes). The reason why this is more reliable is that there can be more than one bank code for a BIC, each with individual values. Also this function avoids the need to do a round trip (if you only have a bank code to start with), which might lead to wrong values (due to the 1:n mapping).

Added test cases for the new function + removed the template text from the license file.